requirements.txt
selenium>=4.18.0
pytest>=8.0.0
pytest-xdist>=3.5.0
webdriver-manager>=4.0.0
allure-pytest>=2.13.0
requests>=2.31.0
pytest-html>=4.1.0
pytest-bdd>=7.0.0
python-dotenv